+++ library/forms/models/MemoryModel.php
@@ -1,7 +1,7 @@
 <?php
-require_once(__DIR__.'/ZB_AbstractModel.php');
+namespace Zerobase\Forms\Models;
 
-class ZB_MemoryModel extends ZB_AbstractModel
+class MemoryModel extends AbstractModel
 {
     protected function storeData($name, $value)
     {
